♻️ refactor(api): add gender schema to client signup and update related imports
