+++ apps/api/src/common/schema/schema.ts
@@ -1,3 +1,5 @@
 import { t } from "elysia";
 
 export const clientSchema = t.Union([t.Literal("web"), t.Literal("mobile")]);
+
+export const genderSchema = t.Optional(t.Union([t.Literal("man"), t.Literal("woman")]));
